Refactor and add additional handlers
