+++ src/authentication.js
@@ -1,7 +1,8 @@
-import { store } from "./store.js";
-import { defaultHandleTokens } from "./cookies.js";
+import { defaultHandleTokens, setTokensFromCookies } from "./tokens.js";
+import { setCookie } from "./cookies.js";
 import { defaultHandleRedirect, getQueryAttr } from "./url.js";
-import { redirectWithPkce } from "./pkce.js";
+import { defaultHandlePkceRequired } from "./pkce.js";
+import { defaultHandleMfaRequired } from "./mfa.js";
 
 // Data specific to the MFA service
 export const authenticationData = {
@@ -11,92 +12,16 @@ export const authenticationData = {
 };
 
 /**
- * Set authenticationData.firstFactors from the authentication object
- * @param {Object} authentication
- * {
- *   firstFactors,
- *   secondFactors
- * }
- * @returns
+ * Set the cookies from a tokens object, and add to the local store.
+ * @param {Object} tokens
  */
-export function setFirstFactors(authentication) {
-  // If we're not initialized, there are no first factors.
-  if (!store.tenantId) {
-    console.warn(
-      "setFirstFactors: tried to set factors without a tenantId set."
-    );
-    return;
+export function setCookiesAndTokens(tokens) {
+  setCookie(tokens.access.value, tokens.access.cookieOptions, "access");
+  setCookie(tokens.id.value, tokens.id.cookieOptions, "id");
+  if (tokens.refresh && tokens.refresh.value) {
+    setCookie(tokens.refresh.value, tokens.refresh.cookieOptions, "refresh");
   }
-  // If we're passed an invalid argument, keep the authentication data as is.
-  if (
-    !authentication ||
-    typeof authentication !== "object" ||
-    !Array.isArray(authentication.firstFactors)
-  ) {
-    console.warn("setFirstFactors: invalid factors passed.");
-    return;
-  }
-  authenticationData.firstFactors = authentication.firstFactors;
-}
-
-/**
- * Check if MFA is required for the ongoing signup or login flow.
- * @returns {Boolean} true if MFA is currently required
- */
-export function isMfaRequired() {
-  return !!authenticationData.firstFactorToken;
-}
-
-/**
- * Update the MFA service state given a response to a signup or login call.
- * Adds secondFactors and firstFactorToken if it is a MFA Required response,
- * removes them if it is a successful signup or login,
- * leaves the service unchanged otherwise.
- * @param {Object} response
- */
-export function handleMfaRequired(response) {
-  if (!response.isMfaRequired) {
-    // If we've logged in or signed up successfully,
-    // clear the MFA service state.
-    if (response.message === "OK") {
-      clearMfa();
-    }
-    return;
-  }
-  authenticationData.secondFactors = response.authentication.secondFactors;
-  authenticationData.firstFactorToken = response.firstFactorToken;
-}
-
-/**
- * If MFA is required, returns a headers object with authorization set to the firstFactorToken.
- * Otherwise, returns an empty object.
- * @returns {Object} a headers object with MFA authorization header set, or empty if MFA is not required
- */
-export function getMfaHeaders() {
-  if (authenticationData.firstFactorToken) {
-    return {
-      authorization: `Bearer ${authenticationData.firstFactorToken}`,
-    };
-  }
-  return {};
-}
-
-/**
- * Clears the current transient state of the MFA service,
- * leaving the tenant's persistent state in place.
- */
-export function clearMfa() {
-  authenticationData.secondFactors = [];
-  authenticationData.firstFactorToken = null;
-}
-
-/**
- * Fully resets the MFA service, including the tenant's persistent state,
- * to it uninitialized state.
- */
-export function resetMfa() {
-  clearMfa();
-  authenticationData.firstFactors = [];
+  setTokensFromCookies();
 }
 
 /**
@@ -104,6 +29,8 @@ export function resetMfa() {
  * @property {Object} data
  * @property {String|Boolean} redirect
  * @property {Function} handleUpstreamResponse
+ * @property {Function} handleMfaRequired
+ * @property {Function} handlePkceRequired
  * @property {Function} handleTokens
  * @property {Function} handleRedirect
  * @returns {Object} data (or redirection)
@@ -112,13 +39,13 @@ export async function handleLoginResponse({
   data,
   redirect,
   handleUpstreamResponse,
+  handleMfaRequired,
+  handlePkceRequired,
   handleTokens,
   handleRedirect,
 }) {
-  const redirectValueIfNotFalse =
+  let redirectValue =
     redirect || getQueryAttr("redirect") || data.redirectTo || "/";
-
-  const redirectValue = redirect === false ? false : redirectValueIfNotFalse;
 
   // Handle upstreamResponse
   if (typeof handleUpstreamResponse === "function") {
@@ -127,7 +54,11 @@ export async function handleLoginResponse({
 
   // Handle "MFA required" response
   if (data.hasOwnProperty("firstFactorToken")) {
-    handleMfaRequired(data);
+    if (typeof handleMfaRequired === "function") {
+      await handleMfaRequired(data.firstFactorToken, data);
+    } else {
+      defaultHandleMfaRequired(data.firstFactorToken, data);
+    }
     return data;
   }
 
@@ -140,21 +71,26 @@ export async function handleLoginResponse({
     }
   }
 
-  // Handle authorizationCode for PKCE
+  // Handle "PKCE required" response
   if (data.hasOwnProperty("authorizationCode")) {
-    if (redirectValueIfNotFalse) {
-      redirectWithPkce(redirectValueIfNotFalse, data.authorizationCode);
-      return data;
-    } else {
+    if (!redirectValue) {
       throw new Error("Missing PKCE redirect url");
+    }
+    if (typeof handlePkceRequired === "function") {
+      await handlePkceRequired(data.authorizationCode, redirectValue, data);
+    } else {
+      defaultHandlePkceRequired(data.authorizationCode, redirectValue, data);
+      return data;
     }
   }
 
   // Handle redirection
-  if (typeof handleRedirect === "function") {
-    await handleRedirect(data.redirectTo, data);
-  } else {
-    defaultHandleRedirect(redirectValue, data);
+  if (data.hasOwnProperty("redirectTo") && redirect !== false) {
+    if (typeof handleRedirect === "function") {
+      await handleRedirect(redirectValue, data);
+    } else {
+      defaultHandleRedirect(redirectValue, data);
+    }
   }
 
   return data;
