+++ test/password-migrate.spec.js
@@ -11,19 +11,19 @@ import {
 } from "./config/utils.js";
 import {
   assertAuthenticationDataMatches,
-  assertNoUser,
   mfaHeaders,
   noMfaHeaders,
   pkceParams,
 } from "./config/assertions.js";
-import { exchange } from "../src/refresh.js";
 import { loginWithPasswordMigrate } from "../src/password.migrate.js";
 import { defaultHandleRedirect } from "../src/url.js";
+import { defaultHandleTokens } from "../src/tokens.js";
 import * as Pkce from "../src/pkce.js";
 
 jest.mock("../src/api.js");
 jest.mock("../src/refresh.js");
 jest.mock("../src/url.js");
+jest.mock("../src/tokens.js");
 jest.mock("../src/pkce.js");
 
 const tenantId = "abcd9876";
@@ -100,18 +100,11 @@ describe("loginWithPasswordMigrate()", () => {
       // Should have returned the proper value
       expect(data).toEqual(mockResponse.data);
 
-      // Should have called exchange() with the API's response
-      expect(exchange).toHaveBeenCalledWith(mockResponse.data);
-
-      // Should have set the user object
-      expect(Userfront.user.email).toEqual(payload.emailOrUsername);
-      expect(Userfront.user.userId).toEqual(idTokenUserDefaults.userId);
+      // Should call defaultHandleTokens correctly
+      expect(defaultHandleTokens).toHaveBeenCalledWith(data.tokens, data);
 
       // Should call defaultHandleRedirect correctly
-      expect(defaultHandleRedirect).toHaveBeenCalledWith(
-        mockResponse.data.redirectTo,
-        mockResponse.data
-      );
+      expect(defaultHandleRedirect).toHaveBeenCalledWith(data.redirectTo, data);
     });
 
     it("should call handleUpstreamResponse before redirecting", async () => {
@@ -119,13 +112,13 @@ describe("loginWithPasswordMigrate()", () => {
       api.post.mockImplementationOnce(() => mockResponse);
 
       // Add a handleUpstreamResponse method
-      const handleFn = jest.fn();
+      const handleUpstreamResponse = jest.fn();
 
       // Call loginWithPasswordMigrate()
       const payload = {
         emailOrUsername: idTokenUserDefaults.email,
         password: "something",
-        handleUpstreamResponse: handleFn,
+        handleUpstreamResponse,
       };
       const data = await loginWithPasswordMigrate(payload);
 
@@ -143,22 +136,15 @@ describe("loginWithPasswordMigrate()", () => {
       // Should have returned the proper value
       expect(data).toEqual(mockResponse.data);
 
-      // Should have called exchange() with the API's response
-      expect(exchange).toHaveBeenCalledWith(mockResponse.data);
-
-      // Should have called handleFn with the upstreamResponse
-      expect(handleFn).toHaveBeenCalledWith(
-        mockResponse.data.upstreamResponse,
-        mockResponse.data
-      );
-
-      // Should have set the user object
-      expect(Userfront.user.email).toEqual(payload.emailOrUsername);
-      expect(Userfront.user.userId).toEqual(idTokenUserDefaults.userId);
+      // Should call defaultHandleTokens correctly
+      expect(defaultHandleTokens).toHaveBeenCalledWith(data.tokens, data);
 
       // Should call defaultHandleRedirect correctly
-      expect(defaultHandleRedirect).toHaveBeenCalledWith(
-        mockResponse.data.redirectTo,
+      expect(defaultHandleRedirect).toHaveBeenCalledWith(data.redirectTo, data);
+
+      // Should have called handleUpstreamResponse with the upstreamResponse
+      expect(handleUpstreamResponse).toHaveBeenCalledWith(
+        mockResponse.data.upstreamResponse,
         mockResponse.data
       );
     });
@@ -196,21 +182,14 @@ describe("loginWithPasswordMigrate()", () => {
         noMfaHeaders
       );
 
-      // Should have called exchange() with the API's response
-      expect(exchange).toHaveBeenCalledWith(mockResponseCopy.data);
-
       // Should have returned the proper value
       expect(data).toEqual(mockResponseCopy.data);
 
-      // Should have set the user object
-      expect(Userfront.user.email).toEqual(payload.email);
-      expect(Userfront.user.userId).toEqual(newAttrs.userId);
+      // Should call defaultHandleTokens correctly
+      expect(defaultHandleTokens).toHaveBeenCalledWith(data.tokens, data);
 
-      // Should call defaultHandleRedirect correctly
-      expect(defaultHandleRedirect).toHaveBeenCalledWith(
-        false,
-        mockResponseCopy.data
-      );
+      // Should not call defaultHandleRedirect
+      expect(defaultHandleRedirect).not.toHaveBeenCalled();
     });
 
     it("should login and redirect to a provided path", async () => {
@@ -221,8 +200,9 @@ describe("loginWithPasswordMigrate()", () => {
         emailOrUsername: idTokenUserDefaults.email,
         password: "something",
       };
-      await loginWithPasswordMigrate({
-        redirect: false,
+      const redirect = "/path";
+      const data = await loginWithPasswordMigrate({
+        redirect,
         ...payload,
       });
 
@@ -236,18 +216,14 @@ describe("loginWithPasswordMigrate()", () => {
         noMfaHeaders
       );
 
-      // Should have called exchange() with the API's response
-      expect(exchange).toHaveBeenCalledWith(mockResponse.data);
+      // Should have returned the proper value
+      expect(data).toEqual(mockResponse.data);
 
-      // Should have set the user object
-      expect(Userfront.user.email).toEqual(payload.emailOrUsername);
-      expect(Userfront.user.userId).toEqual(idTokenUserDefaults.userId);
+      // Should call defaultHandleTokens correctly
+      expect(defaultHandleTokens).toHaveBeenCalledWith(data.tokens, data);
 
       // Should call defaultHandleRedirect correctly
-      expect(defaultHandleRedirect).toHaveBeenCalledWith(
-        false,
-        mockResponse.data
-      );
+      expect(defaultHandleRedirect).toHaveBeenCalledWith(redirect, data);
     });
 
     it("should set the noResetEmail option if provided", async () => {
@@ -323,7 +299,7 @@ describe("loginWithPasswordMigrate()", () => {
       assertAuthenticationDataMatches(mockMfaRequiredResponse);
 
       // Should not have set the user object or redirected
-      assertNoUser(Userfront.user);
+      expect(defaultHandleTokens).not.toHaveBeenCalled();
       expect(defaultHandleRedirect).not.toHaveBeenCalled();
 
       // Should have returned MFA options & firstFactorToken
@@ -388,7 +364,7 @@ describe("loginWithPasswordMigrate()", () => {
           emailOrUsername: idTokenUserDefaults.email,
           password: "something",
         };
-        await loginWithPasswordMigrate(payload);
+        const data = await loginWithPasswordMigrate(payload);
 
         // Should have sent the proper API request
         expect(api.post).toHaveBeenCalledWith(
@@ -400,10 +376,12 @@ describe("loginWithPasswordMigrate()", () => {
           pkceParams("code")
         );
 
-        // Should have requested redirect with the correct params
-        const params = Pkce.redirectWithPkce.mock.lastCall;
-        expect(params[0]).toEqual("my-app:/login");
-        expect(params[1]).toEqual("auth-code");
+        // Should have requested PKCE redirect with the correct params
+        expect(Pkce.defaultHandlePkceRequired).toHaveBeenCalledWith(
+          data.authorizationCode,
+          data.redirectTo,
+          data
+        );
       });
     });
   });
